Replace background for background-color
I think it is a good idea to only set styles for what we want to change. In this file, we only want to change the background color, so it is more precise to use background-color instead of background (shorthand) that also (re)sets default values for -image, -position, -size, -repeat, -origin, -clip, -attachment, and -color.